解决流送的问题

+++ TypeScript/Game/GameController.ts
@@ -4,7 +4,6 @@ import { EditorOperations } from 'ue';
 import { isEditor } from '../Common/Util';
 import { gameConfig } from './Common/GameConfig';
 import { StateComponent } from './Component/StateComponent';
-import { loadTsEntity } from './Entity/Common';
 import { gameContext, IGameController, ITsEntity } from './Interface';
 
 export class GameController implements IGameController {
@@ -20,39 +19,36 @@ export class GameController implements IGameController {
         }
     }
 
-    public LoadGame(): void {
-        const stateManager = gameContext.StateManager;
-        stateManager.Load();
-
-        const entities = gameContext.EntityManager.GetAllEntites();
-        entities.forEach((entity) => {
-            loadTsEntity(entity);
-        });
-    }
-
-    private SaveEntityState(entity: ITsEntity): void {
+    private OnEntityDeregistered(entity: ITsEntity): void {
         const stateComponent = entity.Entity.TryGetComponent(StateComponent);
         if (stateComponent) {
             gameContext.StateManager.SetState(entity.Guid, stateComponent.GenSnapshot());
         }
     }
 
-    private DeleteEntityState(guid: string): void {
-        gameContext.StateManager.DeleteState(guid);
+    private OnEntityRemoved(guid: string): void {
+        const entityMananger = gameContext.EntityManager;
+        const spawnType = entityMananger.GetSpawnType(guid);
+        const destroyType = entityMananger.GetDestoryType(guid);
+        if (spawnType === 'user' && destroyType === 'delete') {
+            gameContext.StateManager.DeleteState(guid);
+        } else if (spawnType === 'streaming' && destroyType === 'delete') {
+            gameContext.StateManager.MarkDelete(guid);
+        }
     }
 
     public SaveGame(): void {
         const entities = gameContext.EntityManager.GetAllEntites();
         entities.forEach((entity) => {
-            this.SaveEntityState(entity);
+            this.OnEntityDeregistered(entity);
         });
         gameContext.StateManager.Save();
     }
 
     public Init(): void {
         const entityMananger = gameContext.EntityManager;
-        entityMananger.EntityRemoved.AddCallback(this.DeleteEntityState.bind(this));
-        entityMananger.EntityDeregistered.AddCallback(this.SaveEntityState.bind(this));
+        entityMananger.EntityRemoved.AddCallback(this.OnEntityRemoved.bind(this));
+        entityMananger.EntityDeregistered.AddCallback(this.OnEntityDeregistered.bind(this));
     }
 
     public Exit(): void {
